+++ .eslintrc.json
@@ -22,6 +22,8 @@
                 "@typescript-eslint"
             ],
             "rules": {
+                "@typescript-eslint/triple-slash-reference": "off",
+                "spaced-comment": ["error", "always", { "markers": ["/"] }],
                 "@typescript-eslint/no-explicit-any": "off",
                 "@typescript-eslint/ban-types": "off",
                 "@typescript-eslint/no-unused-vars": "off",
